--- a/gpustack/policies/candidate_selectors/vllm_resource_fit_selector.py
+++ b/gpustack/policies/candidate_selectors/vllm_resource_fit_selector.py
@@ -194,6 +194,7 @@ class VLLMResourceFitSelector(ScheduleCandidatesSelector):
         self._gpu_count = None
         self._vram_claim = 0
         self._num_attention_heads = None
+        self._message = ""
 
         self._selected_gpu_worker = None
         self._selected_gpu_indexes = []
@@ -252,6 +253,12 @@ class VLLMResourceFitSelector(ScheduleCandidatesSelector):
                 f"({self._gpu_count})."
             )
 
+    def _set_message(self):
+        self._message = "No workers meet the resource requirements. Consider adjusting parameters such as --gpu-memory-utilization (default: 0.9), --max-model-len, or --enforce-eager to lower the resource demands."
+
+    def get_message(self) -> str:
+        return self._message
+
     async def select_candidates(
         self, workers: List[Worker]
     ) -> List[ModelInstanceScheduleCandidate]:
@@ -280,6 +287,7 @@ class VLLMResourceFitSelector(ScheduleCandidatesSelector):
             if candidates:
                 return candidates
 
+        self._set_message()
         return []
 
     async def find_single_worker_single_gpu_full_offloading_candidates(