--- a/tests/controller/test_controller.py
+++ b/tests/controller/test_controller.py
@@ -9,8 +9,8 @@ from gpustack.schemas.models import (
 from gpustack.schemas.workers import WorkerStateEnum
 from gpustack.server.controllers import find_scale_down_candidates
 from tests.fixtures.workers.fixtures import (
-    linux_nvidia_2_4090x2,
-    linux_nvidia_2_4080x2,
+    linux_nvidia_19_4090_24gx2,
+    linux_nvidia_2_4080_16gx2,
     linux_cpu_1,
 )
 
@@ -21,11 +21,11 @@ from tests.utils.model import new_model, new_model_instance
 
 @pytest.mark.asyncio
 async def test_find_scale_down_candidates():
-    w1 = linux_nvidia_2_4090x2()
+    w1 = linux_nvidia_19_4090_24gx2()
     w1.state = WorkerStateEnum.NOT_READY
     workers = [
         w1,
-        linux_nvidia_2_4080x2(),
+        linux_nvidia_2_4080_16gx2(),
         linux_cpu_1(),
     ]