--- a/tests/fixtures/workers/linux_nvidia_11_V100_32gx2.json
+++ b/tests/fixtures/workers/linux_nvidia_11_V100_32gx2.json
@@ -1,0 +1,65 @@
+{
+    "name": "host01-v100",
+    "hostname": "host01-v100",
+    "ip": "192.168.50.21",
+    "labels": {
+        "os": "Linux"
+    },
+    "state": "ready",
+    "system_reserved": {
+        "memory": 2147483648,
+        "gpu_memory": 0
+    },
+    "status": {
+        "memory": {
+            "total": 1610612736000,
+            "utilization_rate": 5.905919020930682,
+            "is_unified_memory": false,
+            "used": 3967795200,
+            "allocated": 0
+        },
+        "gpu_devices": [
+            {
+                "uuid": "GPU-e41f9601-2307-3af8-98c3-f25afe04c263",
+                "name": "V100",
+                "vendor": "NVIDIA",
+                "index": 0,
+                "core": {
+                    "total": 10240,
+                    "utilization_rate": 0.0
+                },
+                "memory": {
+                    "total": 34359738368,
+                    "utilization_rate": 2.500228993649243,
+                    "is_unified_memory": false,
+                    "used": 429326336,
+                    "allocated": 0
+                },
+                "temperature": 36.0,
+                "type": "cuda"
+            },
+            {
+                "uuid": "GPU-e41f9601-2307-3af8-98c3-f25afe04c264",
+                "name": "V100",
+                "vendor": "NVIDIA",
+                "index": 1,
+                "core": {
+                    "total": 10240,
+                    "utilization_rate": 0.0
+                },
+                "memory": {
+                    "total": 34359738368,
+                    "utilization_rate": 3.6070316316560818,
+                    "is_unified_memory": false,
+                    "used": 619380736,
+                    "allocated": null
+                },
+                "temperature": 38.0,
+                "type": "cuda"
+            }
+        ]
+    },
+    "id": 17,
+    "created_at": "2024-07-05T02:17:42",
+    "updated_at": "2024-07-05T10:36:21"
+}