--- a/tests/utils/model.py
+++ b/tests/utils/model.py
@@ -44,6 +44,7 @@ def new_model(
         "distributed_inference_across_workers", True
     )
     gpu_selector = kargs.get("gpu_selector", None)
+    backend_parameters = kargs.get("backend_parameters")
 
     source = SourceEnum.OLLAMA_LIBRARY
     huggingface_filename = None
@@ -66,4 +67,5 @@ def new_model(
         cpu_offloading=cpu_offloading,
         distributed_inference_across_workers=distributed_inference_across_workers,
         gpu_selector=gpu_selector,
+        backend_parameters=backend_parameters,
     )