… gpustack/policies/candidate_selectors/gguf_resource_fit_selector.py
… tests/fixtures/estimates/deepseek_r1_distill_qwen_32b_bf16_partial_offload.json
… tests/fixtures/estimates/deepseek_r1_distill_qwen_32b_bf16_partial_offload_split_1main_1rpc.json
… tests/fixtures/estimates/deepseek_r1_q4_k_m_full_offload.json
… tests/fixtures/estimates/deepseek_r1_q4_k_m_partial_offload_split_1main_1rpc.json
… tests/fixtures/estimates/fixtures.py
… tests/fixtures/estimates/llama3_70b_full_offload.json
… tests/fixtures/estimates/llama3_70b_full_offload_split_2_4080.json
… tests/fixtures/estimates/llama3_70b_full_offload_split_2_4090.json
… tests/fixtures/estimates/llama3_70b_full_offload_split_2_4090_2.json
… tests/fixtures/estimates/llama3_70b_full_offload_split_2_4090_3.json
… tests/fixtures/estimates/llama3_70b_full_offload_split_3_4080.json
… tests/fixtures/estimates/llama3_70b_full_offload_split_3_4080_2.json
… tests/fixtures/estimates/llama3_70b_full_offload_split_3_4080_3.json
… tests/fixtures/estimates/llama3_70b_full_offload_split_4_4080.json
… tests/fixtures/estimates/llama3_70b_partial_offload_split_2_4090.json
… tests/fixtures/estimates/llama3_70b_partial_offload_split_3_4080.json
… tests/fixtures/estimates/llama3_70b_partial_offload_split_3_4080_2.json
… tests/fixtures/estimates/llama3_70b_partial_offload_split_3_4080_3.json
… tests/fixtures/estimates/llama3_70b_partial_offload_split_3_4080_4.json
… tests/fixtures/estimates/llama3_8b_full_offload.json
… tests/fixtures/estimates/llama3_8b_partial_offload_split_1main_1rpc.json
… tests/fixtures/estimates/unsloth_DeepSeek-R1-GGUF_DeepSeek-R1-BF16/remote/brief/6/668b33ce3725cb60
… tests/fixtures/workers/linux_nvidia_13_A100_80gx8.json
… tests/fixtures/workers/linux_nvidia_15_4080_16gx8.json
… tests/fixtures/workers/linux_nvidia_16_5000_16gx8.json
… tests/fixtures/workers/linux_nvidia_17_4090_24gx8.json
… tests/fixtures/workers/linux_nvidia_18_4090_24gx4_4080_16gx4.json
… tests/fixtures/workers/linux_nvidia_19_4090_24gx2.json
… tests/fixtures/workers/linux_nvidia_1_4090_24gx1.json