## Issue
https://github.com/gpustack/gpustack/issues/1312

## Solution
### Cache
- Cache for static reusable data
  - full offload resource claim
  - disable offload resource claim
  - partial offload resource claim(without tensor split, rpc)
  - the minimum required vram of offload a layer
- Cache for dynamic data
  - cached partial offload data when calculate resource claim for single worker multiple gpus and multiple workers multiple gpus. For example: worker with 4 same gpus, gpu combinations(0, 1),(0,2),(0,3),(1,2),(2,3) could use the same cached resource claim.
  - cached partial offload for multi worker multi gpus

### Prune
- Full offload: Sort workers/gpus allocatable resource before scheduled, skip subsequent workers/gpus evaluate since they have less vram
- Full offload: Prune workers and GPUs that allocatable resource is less then full offload resource to single gpu
   - Base on vram claim: full offload to single gpu < full offload to single worker multi gpu < full offload to multi worker multi gpu 
- Partial offload: Prune workers and GPUs that cannot offload even a single layer
  - For single worker multiple gpus deployment, prune workers and GPUs that cannot offload even a single layer based on the minimum VRAM required to offload one layer before generating GPU combinations.
  - For multi worker multiple gpus deployment, prune GPUs that cannot offload even a single layer based on the minimum VRAM required to offload one layer for rpc before generating GPU combinations.
  - For multi worker multiple gpus deployment, initialize the max offload layers base on the cached partial offload resource claim. In future evaluations of combinations, only assess those with the number of offloaded layers greater than the max layers. Continuously update the maximum number of offloaded layers.

### Other
- Enhanced logging

## Compare the time before and after automatic scheduling
Condition: deploy unsloth_DeepSeek-R1-GGUF_DeepSeek-R1-BF16

### Case 1:
- Workers

| Device        | Device        | Device        |
|---------------|---------------|---------------|
| 3090_24gx8    | V100_32gx2    | A40_48gx2     |
| A100_80gx8    | A100_40gx2    |               |


- Expected result: multiple workers multi gpus patial offload
- Laytency:

| case | before schedule time | after schedule time |after schedule time without cache |
| ---- | -------------------- | ------------------- |-----|
| Auto | >30m                 | ~3.5s               |~5s |

### Case 2:
- Workers

| Device        | Device        | Device        |
|---------------|---------------|---------------|
| 4080_16gx2    | 4080_16gx4    | 4090_24gx2    |
| V100_32gx2    | A40_48gx2     | A100_40gx2    |
| 4080_16gx8    | 3090_24gx8    | 3090_24gx8    |
| 3090_24gx8    | m2_24g        | m1pro_21g     |

- Expected result: single patial offload
- Laytency:

| case | before schedule time | after schedule time | after schedule time without cache |
| ---- | -------------------- | ------------------- | -----|
| Auto | >30m                 | ~3.6s               | ~5s |

### Case 3:
- Workers

| Device               | Device               | Device               | Device               |
|----------------------|----------------------|----------------------|----------------------|
| 4090_24gx1           | 4080_16gx2           | 4080_16gx4           | 4090_24gx1           |
| 3090_24gx8           | 3090_24gx8           | 3090_24gx8           | 4080_16gx8           |
| 5000_16gx8           | 4090_24gx8           | 4090_24gx4_4080_16gx4| 4090_24gx2           |
| 3080_12gx8           | 4090_24gx4_3060_12gx4| 7800_16gx1           | m2_24g               |
| m1pro_21g            | cpu_1                | cpu_2                | cpu_3                |


- Expected result:cpu offload
- Laytency:

| case | before schedule time | after schedule time |after schedule time without cache |
| ---- | -------------------- | ------------------- |-----|
| Auto | > 30min              | ~1s               | ~1s|